# schedula/utils/form/server/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Schedula API
  version: 1.0.0
  description: OpenAPI spec generato dal codice server (Flask). Alcuni endpoint sono
    condizionati da feature-flag di configurazione.
servers:
- url: /
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
    authToken:
      type: apiKey
      in: header
      name: Authentication-Token
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
security:
- bearerAuth: []
- cookieAuth: []
- authToken: []
paths:
...
  /groups:
    get:
      tags:
      - Groups
      summary: List my groups
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - &id002
        name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 200
      - &id003
        name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
      - &id004
        name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
      - &id005
        name: page_size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 200
      - &id007
        name: sort
        in: query
        required: false
        schema:
          type: string
        description: 'Campo e direzione, es: created_at:-1'
      - name: admin_only
        in: query
        required: false
        schema:
          type: boolean
      - name: type
        in: query
        required: false
        schema:
          type: string
      - name: name
        in: query
        required: false
        schema:
          type: string
    post:
      tags:
      - Groups
      summary: Create group
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  group:
                    type: object
                    additionalProperties: true
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                type:
                  type: string
              required:
              - name
  /groups/{gid}:
    get:
      tags:
      - Groups
      summary: Get group
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - &id001
        name: gid
        in: path
        required: true
        schema:
          type: string
    put:
      tags:
      - Groups
      summary: Update group (PUT)
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                type:
                  type: string
              required:
              - name
    patch:
      tags:
      - Groups
      summary: Update group (PATCH)
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                type:
                  type: string
  /groups/{gid}/memberships:
    patch:
      tags:
      - Groups
      summary: Edit group memberships (batch)
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                add_members:
                  type: array
                  items:
                    type: string
                remove_members:
                  type: array
                  items:
                    type: string
                promote_admins:
                  type: array
                  items:
                    type: string
                demote_admins:
                  type: array
                  items:
                    type: string
...